* Source/NSArray.m (-objectsAtIndexes:): Fix comparision of unsigned variable with NSNotFound.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@34389 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSArray.m
@@ -925,7 +925,7 @@ static SEL	rlSel;
   //FIXME: probably slow!
   NSMutableArray *group = [NSMutableArray arrayWithCapacity: [indexes count]];
 
-  unsigned i = [indexes firstIndex];
+  NSUInteger i = [indexes firstIndex];
   while (i != NSNotFound)
     {
       [group addObject: [self objectAtIndex: i]];
